Move the functionality to CreationHelper.
